pyproject.toml
[project]
name = "bm25s"
version = "0.1.0"
description = "BM25 model in token and semantic domains"
license = "GPL-3.0-or-later"
authors = [
    {name = "Igor Y. Khomyakov", email = "ikh@ikhsoftware.com"},
]
dependencies = [
    "pyyaml",
    "numpy",
    "tqdm",
    "transformers",
]
requires-python = ">=3.11"
readme = "README.md"


[tool.pdm.dev-dependencies]
lint = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.1",
    "pylint>=3.0.2",
]
